requirements.txt
Flask==2.2.2
Werkzeug==2.3.4
authlib
python-dotenv